watchdog_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-watchdog-env"
version = "0.1.0"
description = "WatchDog: RL environment for training AI oversight agents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]
plugins = [
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
]

[project.scripts]
server = "watchdog_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["watchdog_env", "watchdog_env.server", "watchdog_env.plugins", "watchdog_env.plugins.cicero", "watchdog_env.plugins.avalon"]
package-dir = { "watchdog_env" = ".", "watchdog_env.server" = "server", "watchdog_env.plugins" = "plugins", "watchdog_env.plugins.cicero" = "plugins/cicero", "watchdog_env.plugins.avalon" = "plugins/avalon" }
